Add validator dashboard, qr scanner, validate page

+++ frontend/src/app/admin-dashboard/manage/manage-events/manage-events.component.ts
@@ -13,13 +13,15 @@ import {
 import {MatMenu, MatMenuItem, MatMenuTrigger} from "@angular/material/menu";
 import {MatPaginator} from "@angular/material/paginator";
 import {MatSort, MatSortHeader} from "@angular/material/sort";
-import {NgForOf} from "@angular/common";
+import {NgForOf, NgIf} from "@angular/common";
 import {MatDialog} from "@angular/material/dialog";
 import {MatSnackBar} from "@angular/material/snack-bar";
 import {ManageLocationsService} from "../manage-locations/manage-locations.service";
 import {ManageEventsService} from "./manage-events.service";
 import {LocationNamePipe} from "../../../util/pipes/location-name.pipe";
 import {AddEditEventComponent} from "./popup/add-edit-event/add-edit-event.component";
+import {DeleteComponent} from "../shared/delete/delete.component";
+import {LoadingComponent} from "../../../shared/loading/loading.component";
 
 @Component({
   selector: 'app-manage-events',
@@ -44,12 +46,16 @@ import {AddEditEventComponent} from "./popup/add-edit-event/add-edit-event.compo
     NgForOf,
     MatHeaderCellDef,
     MatMenuTrigger,
-    LocationNamePipe
+    LocationNamePipe,
+    NgIf,
+    LoadingComponent
   ],
   templateUrl: './manage-events.component.html',
   styleUrl: './manage-events.component.scss'
 })
 export class ManageEventsComponent implements OnInit{
+  loading: boolean = true;
+
   dataSource = new MatTableDataSource<any>();
   @ViewChild(MatPaginator) paginator: MatPaginator | undefined;
   @ViewChild(MatSort) sort: MatSort | undefined;
@@ -81,9 +87,12 @@ export class ManageEventsComponent implements OnInit{
         this.dataSource.paginator = this.paginator!;
       },
       error: (error: any) => {
-        this.snackBar.open('Error fetching users', 'Close', {
+        this.snackBar.open(error.error, 'Close', {
           duration: 3000
         });
+      },
+      complete: () => {
+        this.loading = false;
       }
     });
   }
@@ -163,14 +172,20 @@ export class ManageEventsComponent implements OnInit{
   }
 
   onDeleteClick() {
-    this.manageEventsService.deleteEvent(this.rowData.id).subscribe({
-      next: () => {
-        this.dataSource.data = this.dataSource.data.filter((e: any) => e.id !== this.rowData.id);
-        this.snackBar.open('Eveniment sters cu succes', 'Inchide', {duration: 3000});
-      },
-      error: (error: any) => {
-        this.snackBar.open('A aparut o eroare la stergerea evenimentului', 'Inchide', {duration: 3000});
-      }
+    let dialogRef = this.dialog.open(DeleteComponent, {
+      width: '30%'
+    });
+    dialogRef.afterClosed().subscribe((result: any) => {
+      if(result) {
+        this.manageEventsService.deleteEvent(this.rowData.id).subscribe({
+          next: () => {
+            this.dataSource.data = this.dataSource.data.filter((e: any) => e.id !== this.rowData.id);
+            this.snackBar.open('Eveniment sters cu succes', 'Inchide', {duration: 3000});
+          },
+          error: (error: any) => {
+            this.snackBar.open('A aparut o eroare la stergerea evenimentului', 'Inchide', {duration: 3000});
+          }
+        });      }
     });
   }
 }
